refactor: TreeNode 相关类到文件夹下
