+++ src/BootstrapBlazor/Extensions/ITableTreeItemExtensions.cs
@@ -2,55 +2,12 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 // Website: https://www.blazor.zone or https://argozhang.github.io/
 
-namespace BootstrapBlazor.Components;
-
-/// <summary>
-/// Table Tree 资料结构接口
-/// </summary>
-/// <typeparam name="TItem"></typeparam>
-/// <remarks>
-/// <code>接口实作方式:<br/> class <typeparamref name="TItem"/> : <see cref="ITableTreeItem{TItem}"/> </code>
-/// </remarks>
-public interface ITableTreeItem<TItem>
-{
-    /// <summary>
-    /// 实例化接口
-    /// </summary>
-    /// <typeparam name="TNew"></typeparam>
-    /// <param name="item"></param>
-    /// <returns></returns>
-    public static ITableTreeItem<TNew> New<TNew>(TNew item) where TNew : class
-    {
-        if (item is ITableTreeItem<TNew> ret)
-        {
-            return ret;
-        }
-        else
-        {
-            return new TableTreeNode<TNew>(item);
-        }
-    }
-
-    /// <summary>
-    /// 获得 子节点集合
-    /// </summary>
-    public IEnumerable<ITableTreeItem<TItem>>? Children { get; }
-
-    /// <summary>
-    /// 设置 子节点集合
-    /// </summary>
-    public void SetChildren(IEnumerable<TItem> items);
-
-    /// <summary>
-    /// 获得/设置 是否展开
-    /// </summary>
-    public bool IsExpand { get; set; }
-}
+namespace BootstrapBlazor.Extensions;
 
 /// <summary>
 /// 
 /// </summary>
-public static class ITableTreeItemExtensions
+internal static class ITableTreeItemExtensions
 {
     /// <summary>
     /// 尝试在全部树状结构 <paramref name="items"/> 中寻找指定 <paramref name="target"/>
@@ -79,7 +36,7 @@ public static class ITableTreeItemExtensions
     /// <remarks>采广度优先搜寻</remarks>
     public static ITableTreeItem<TItem>? Find<TItem>(this IEnumerable<ITableTreeItem<TItem>> items, TItem target, Func<TItem, TItem, bool>? equals = null) where TItem : class
     {
-        return Find(items, target, out _, equals);
+        return items.Find(target, out _, equals);
     }
 
     /// <summary>
@@ -178,4 +135,3 @@ public static class ITableTreeItemExtensions
         }
     }
 }
-
